The loading of the world stops for some unknown reason. The client is waiting for something from the server. I can't understand why it doesn't work. Fixed relogin; Updatet some packets; Updated table Characters & Housings int -> bigint for account_Id;

+++ AAEmu.Game/Core/Packets/G2L/GLRequestInfoPacket.cs
@@ -7,11 +7,11 @@ namespace AAEmu.Game.Core.Packets.G2L
 {
     public class GLRequestInfoPacket : LoginPacket
     {
-        private readonly ulong _connectionId;
+        private readonly uint _connectionId;
         private readonly uint _requestId;
         private readonly List<LoginCharacterInfo> _characters;
 
-        public GLRequestInfoPacket(ulong connectionId, uint requestId, List<LoginCharacterInfo> characters) : base(0x03)
+        public GLRequestInfoPacket(uint connectionId, uint requestId, List<LoginCharacterInfo> characters) : base(0x03)
         {
             _connectionId = connectionId;
             _requestId = requestId;
